Correção de calendário e posicionamento

+++ application/src/java/bpmn/artifact/Artifact.java
@@ -23,7 +23,7 @@ public class Artifact extends Element{
    
     }
 
-    public Artifact(String id, int x, int y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
+    public Artifact(String id, float x, float y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
         super(id, x, y, description, elements, transitions, vinculos);
         
         this.setType("BPMNData");
