+++ application/src/java/bpmn/artifact/DataObject.java
@@ -20,7 +20,7 @@ public class DataObject extends Artifact{
     public DataObject() {
     }
 
-    public DataObject(String id, int x, int y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
+    public DataObject(String id, float x, float y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
         super(id, x, y, description, elements, transitions, vinculos);
         this.setName("data-object");
     }
