+++ application/src/java/bpmn/artifact/DataStore.java
@@ -20,7 +20,7 @@ public class DataStore extends Artifact{
     public DataStore() {
     }
 
-    public DataStore(String id, int x, int y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
+    public DataStore(String id, float x, float y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
         super(id, x, y, description, elements, transitions, vinculos);
         this.setName("data-store");
     }
