+++ application/src/java/bpmn/event/EndEvent.java
@@ -20,7 +20,7 @@ public class EndEvent extends Event{
     public EndEvent() {
     }
 
-    public EndEvent(String id, int x, int y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
+    public EndEvent(String id, float x, float y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
         super(id, x, y, description, elements, transitions, vinculos);
         setName("end-event-none");
     }
