+++ application/src/java/bpmn/event/EndEventCancel.java
@@ -20,7 +20,7 @@ public class EndEventCancel extends EndEvent{
     public EndEventCancel() {
     }
 
-    public EndEventCancel(String id, int x, int y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
+    public EndEventCancel(String id, float x, float y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
         super(id, x, y, description, elements, transitions, vinculos);
         setName("end-event-cancel");
     }
