+++ application/src/java/bpmn/event/StartEvent.java
@@ -21,7 +21,7 @@ public class StartEvent extends Event{
         
     }
 
-    public StartEvent(String id, int x, int y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
+    public StartEvent(String id, float x, float y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
         super(id, x, y, description, elements, transitions, vinculos);
         setName("start-event-none");
     }
