+++ application/src/java/bpmn/event/StartEventCompensation.java
@@ -20,7 +20,7 @@ public class StartEventCompensation extends StartEvent {
     public StartEventCompensation() {
     }
 
-    public StartEventCompensation(String id, int x, int y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
+    public StartEventCompensation(String id, float x, float y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
         super(id, x, y, description, elements, transitions, vinculos);
         setName("start-event-compensation");
     }
