+++ application/src/java/bpmn/event/StartEventMultiple.java
@@ -20,7 +20,7 @@ public class StartEventMultiple extends StartEvent{
     public StartEventMultiple() {
     }
 
-    public StartEventMultiple(String id, int x, int y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
+    public StartEventMultiple(String id, float x, float y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
         super(id, x, y, description, elements, transitions, vinculos);
         setName("start-event-multiple");
     }
