+++ application/src/java/bpmn/gateway/Gateway.java
@@ -20,7 +20,7 @@ public class Gateway extends Element{
     public Gateway() {
     }
 
-    public Gateway(String id, int x, int y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
+    public Gateway(String id, float x, float y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
         super(id, x, y, description, elements, transitions, vinculos);
         this.setType("BPMNGateway");
         setName("gateway-none");
