+++ application/src/java/bpmn/gateway/GatewayComplex.java
@@ -20,7 +20,7 @@ public class GatewayComplex extends Gateway{
     public GatewayComplex() {
     }
 
-    public GatewayComplex(String id, int x, int y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
+    public GatewayComplex(String id, float x, float y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
         super(id, x, y, description, elements, transitions, vinculos);
         setName("gateway-complex");
     }
