+++ application/src/java/bpmn/gateway/GatewayEventBased.java
@@ -20,7 +20,7 @@ public class GatewayEventBased extends Gateway{
     public GatewayEventBased() {
     }
 
-    public GatewayEventBased(String id, int x, int y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
+    public GatewayEventBased(String id, float x, float y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
         super(id, x, y, description, elements, transitions, vinculos);
         setName("gateway-eventbased");
     }
