+++ application/src/java/bpmn/gateway/GatewayOR.java
@@ -20,7 +20,7 @@ public class GatewayOR extends Gateway {
     public GatewayOR() {
     }
 
-    public GatewayOR(String id, int x, int y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
+    public GatewayOR(String id, float x, float y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
         super(id, x, y, description, elements, transitions, vinculos);
         setName("gateway-or");
     }
