+++ application/src/java/bpmn/task/ScriptTask.java
@@ -20,7 +20,7 @@ public class ScriptTask extends Task{
     public ScriptTask() {
     }
 
-    public ScriptTask(String id, int x, int y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
+    public ScriptTask(String id, float x, float y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
         super(id, x, y, description, elements, transitions, vinculos);
         setName("script-task");
     }
