+++ application/src/java/bpmn/task/SendTask.java
@@ -20,7 +20,7 @@ public class SendTask extends Task{
     public SendTask() {
     }
 
-    public SendTask(String id, int x, int y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
+    public SendTask(String id, float x, float y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
         super(id, x, y, description, elements, transitions, vinculos);
         setName("send-task");
     }
