+++ application/src/java/bpmn/task/UserTask.java
@@ -20,7 +20,7 @@ public class UserTask extends Task{
     public UserTask() {
     }
 
-    public UserTask(String id, int x, int y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
+    public UserTask(String id, float x, float y, String description, ArrayList<Element> elements, ArrayList<ConnectionObject> transitions, ArrayList<String> vinculos) {
         super(id, x, y, description, elements, transitions, vinculos);
         setName("user-task");
     }
